requirements.txt
requests~=2.25.1
eve~=1.1.4
dnspython
Flask~=1.1.2
pymongo~=3.11.2
gunicorn==20.0.4
firebase-admin==4.5.2
google-cloud-pubsub==2.3.0